requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5
langchain-core>=0.1.10
langchain-text-splitters>=0.0.1
langchain-chroma>=0.0.4
openai>=1.3.0
tiktoken>=0.5.1
python-dotenv>=1.0.0
streamlit>=1.30.0
streamlit-extras>=0.3.5
tqdm>=4.66.1
chromadb>=0.4.22
lxml>=4.9.3
markdownify>=0.11.6
atlassian-python-api>=3.41.2
requests>=2.31.0
beautifulsoup4>=4.12.2

# 추가 유틸리티 패키지
pandas>=2.0.0
numpy>=1.24.0
tenacity>=8.2.0

# 선택적 패키지 (필요에 따라 주석 해제)
# torch>=2.0.0
# sentence-transformers>=2.2.2  # 로컬 임베딩 모델을 위한 패키지
